Allow to pass a block_device and opf to bio_alloc
Switch the arguments of bio_alloc() depending on RHEL releases so that
segment buffer code passes a block_device and opf to bio_alloc() for
newer kernels (RHEL_RELEASE >= 120).

This also refactors the segbuf write code to pass the op to bio_alloc
instead of setting it just before the submission.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>