Allow switching between new and old blkdev_{get_by_*,put}()
In kernel 6.5, an argument was added to blkdev_get_by_*() that takes a
pointer to a blk_holder_ops structure for exclusive claims, and
blkdev_put() has been changed not to use FMODE_EXCL flag.

Define wrapper macros so that these function interfaces can be switched
depending on the kernel version, and use these macros in nilfs_mount().

For RHEL 9 clone kernels, the new interfaces are applied for
RHEL_RELEASE >= 370.

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@gmail.com>